Fix issue #97
