+++ jtwig-core/src/main/java/com/lyncode/jtwig/addons/spaceless/Spaceless.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.lyncode.jtwig.test.addons.spaceless;
+package com.lyncode.jtwig.addons.spaceless;
 
 import com.lyncode.jtwig.JtwigContext;
 import com.lyncode.jtwig.exception.RenderException;
